init(main): starting scientific initiation project
This commit marks the beginning of the project's development of the main
file. The initial idea is to take individual images to test the
compression levels that can be applied to any image in the dataset.
Next steps are:
- [ ] Go through the directories to compress all images in the dataset;
- [ ] Collect the name of the file being processed to include in the
output file;
- [ ] Remove colors from images;
- [ ] Reduce image colors;
- [ ] Reduce image size;
- [ ] Begin model development for training and validation;
- [ ] Create graphs to identify significant loss in accuracy and
F1-score;
- [ ] Develop an algorithm that determines the compression level that
the model can identify without significant loss of accuracy and apply
lossy compression using another algorithm.